@@ -22,7 +22,6 @@
   "dependencies": {
     "@types/chalk": "^2.2.0",
     "axios": "0.19.2",
-    "axios-retry": "^3.1.9",
     "chalk": "^4.1.0",
     "cookie-parse": "^0.4.0",
     "coroutine": "^0.3.1",
@@ -38,6 +37,7 @@
     "node-rsa": "^1.0.8",
     "parse-headers": "^2.0.3",
     "qs": "^6.9.4",
+    "retry-axios": "^2.4.0",
     "rsuite": "^4.0.4",
     "superagent": "^5.2.2",
     "superagent-retry-delay": "^2.6.2",
@@ -49,6 +49,6 @@
     "babel-preset-rsuite": "^4.0.0",
     "ts-node": "^9.1.1",
     "tsc-alias": "^1.2.6",
-    "typescript": "^3.6.4"
+    "typescript": "^3.7.4"
   }
 }
