@@ -1,18 +1,31 @@
 import axios from 'axios'
 import {NuxtAxiosInstance} from "../type";
-const axiosRetry = require('axios-retry');
+import * as rax from 'retry-axios';
+import {log4js} from "@/log";
 const headers = {
     'Connection': 'keep-alive',
         'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8',
         'User-Agent': '',
         'Accept-Encoding': 'gzip,deflate',
 }
-
+const logger = log4js.getLogger(`HTTP`)
 // @ts-ignore
 const $axios :NuxtAxiosInstance = axios.create({ headers , withCredentials: true})
-axiosRetry($axios, { retries: 6 , retryDelay: (retryCount :number) => {
-    return retryCount * 2500
-    }});
+
+$axios.defaults.raxConfig = {
+    instance: $axios,
+    retry: 5,
+    retryDelay: 1000,
+    httpMethodsToRetry: ['GET', 'HEAD', 'OPTIONS', 'DELETE', 'PUT', 'POST'],
+    noResponseRetries: 5,
+    onRetryAttempt: err => {
+        const cfg = rax.getConfig(err);
+        logger.error(`Retry attempt #${cfg && cfg.currentRetryAttempt}`);
+    }
+}
+
+rax.attach($axios)
+
 for (const method of ['request', 'delete', 'get', 'head', 'options', 'post', 'put', 'patch']) {
     // @ts-ignore
     $axios['$' + method] = function () { return this[method].apply(this, arguments).then(res => res && (res.data.resHeaders = res.headers) && res.data) }
