@@ -1,8 +1,9 @@
 export interface VideoPart {
-    title: string,
-    desc: string,
-    path?: string,
-    filename?: string,
-    server_file_name?: string,
-    fileSize?: number
+    title: string;
+    desc: string;
+    path?: string | undefined;
+    filename?: string;
+    server_file_name?: string;
+    fileSize?: number;
+    isFail?: boolean;
 }
