@@ -7,7 +7,7 @@ export const testRoomTypeArr = (roomType: string) => {
   if (RoomTypeArr.some((type) => type === roomType)) return roomType;
   else return "error";
 };
-export const getRoomArrInfo = (roomObj: { [key: string]: { roomUrl: string, tags: string[], tid: number, deleteLocalFile: boolean, uploadLocalFile: boolean } }[]): StreamInfo[] => {
+export const getRoomArrInfo = (roomObj: { [key: string]: { denyTime: number, copyright:number, dynamic: string, source: string, desc: string, roomUrl: string, tags: string[], tid: number, deleteLocalFile: boolean, uploadLocalFile: boolean, templateTitle: string } }[]): StreamInfo[] => {
   let roomInfoArr = [];
   for (let roomInfo of roomObj) {
     for (let key in roomInfo) {
@@ -17,8 +17,12 @@ export const getRoomArrInfo = (roomObj: { [key: string]: { roomUrl: string, tags
       let roomTid = roomInfo[key].tid
       let deleteLocalFile = roomInfo[key].deleteLocalFile
       let uploadLocalFile = roomInfo[key].uploadLocalFile
-      // const repObj: any = /\.(\S+)\./.exec(roomLink);
-      // let roomType = testRoomTypeArr(repObj[1]);
+      let templateTitle = roomInfo[key].templateTitle
+      let desc = roomInfo[key].desc
+      let source = roomInfo[key].source
+      let dynamic = roomInfo[key].dynamic
+      let copyright = roomInfo[key].copyright
+      let denyTime = roomInfo[key].denyTime
       roomInfoArr.push({
         roomName,
         roomLink,
@@ -26,7 +30,13 @@ export const getRoomArrInfo = (roomObj: { [key: string]: { roomUrl: string, tags
         streamUrl: "",
         roomTid,
         deleteLocalFile,
-        uploadLocalFile
+        uploadLocalFile,
+        templateTitle,
+        desc,
+        source,
+        dynamic,
+        copyright,
+        denyTime
       });
     }
   }
